refactor(umbp): generic component metrics for backend-agnostic storage and transfer
Replace the per-medium counter mechanism (MediumBackend::Counters(), keyed
by an SSD-specific header and a hand-copied dashboard JSON) with a MetricSource
/ MetricPublisher abstraction: an InstrumentedBackend decorator derives
generic ops/bytes/latency series for any storage backend from the
MediumBackend interface, and CompositeTransferEngine now measures its own
per-engine bytes, plans, failures, and in-flight time at the one point that
knows which transport carried which plan. A medium or transfer engine no
longer needs its own metrics code or its own dashboard panel — tier/backend/
engine become labels instead of separate metric identifiers.

Dashboards follow that split rather than collapsing into one. Only the
backend-specific dashboard is merged: umbp_ssd_tier.json becomes
umbp_backends.json, whose panels group by tier and engine so every medium and
every transport shares them and a new one appears with no dashboard change.
The four backend-agnostic dashboards — client data rate/bandwidth, RPC and
call rates, master client RPC latency, and the external-KV index — measure the
client, the master and the KV index, look the same whatever medium is
underneath, and are left exactly as they were. All five stay plain Grafana
JSON, editable in the UI and exported back.

Also fix MetricsServer::addCounter truncating fractional deltas (e.g.
*_seconds_total) by taking uint64_t deltas instead of double, and switch
value serialization to print exact integers instead of ostream's
six-significant-digit default.

+++ include/mori/metrics/prometheus_metrics_server.hpp
@@ -102,9 +102,16 @@ class MetricsServer {
   void setGauge(std::string_view name, std::string_view help, const Labels& labels, double value);
 
   // Add `delta` to a monotonically-increasing counter (default delta = 1).
-  void addCounter(std::string_view name, std::string_view help, uint64_t delta = 1);
+  //
+  // `delta` is a double because a Prometheus counter is a double, and because
+  // not every counter counts whole things: a *_seconds_total accumulates
+  // fractions of a second, and taking a uint64 here silently floored every one
+  // of them to zero.  Whole-number counters are unaffected — see FormatValue in
+  // the .cpp, which serialises an integral double exactly rather than through
+  // the six-significant-digit default that would round a byte total.
+  void addCounter(std::string_view name, std::string_view help, double delta = 1);
   void addCounter(std::string_view name, std::string_view help, const Labels& labels,
-                  uint64_t delta = 1);
+                  double delta = 1);
 
   // Record one histogram observation.
   // `bounds` is an ascending list of finite upper-bound values; the implicit
@@ -149,12 +156,12 @@ class MetricsServer {
 
   struct CounterEntry {
     std::string help;
-    uint64_t value{0};
+    double value{0.0};
   };
 
   struct LabeledCounterFamily {
     std::string help;
-    std::map<std::string, uint64_t> series;
+    std::map<std::string, double> series;
   };
 
   struct LabeledGaugeFamily {
